pyproject.toml
[project]
name = "todoist-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "todoist-api-python>=3.0.1",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.7",
    "types-requests>=2.32.0.20250328",
]

[tool.ruff]
select = [
    "E", "W", "F",
    "B", "C", "N",
    "UP",
    "I",
]

[tool.mypy]
exclude = 'venv|.venv|env|.env'
ignore_missing_imports = true
strict_optional = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
follow_imports = "silent"
